Chips script created to help identify each chip whether it red or blue.  Chips have SetChip function that takes 1 or 0 and a material. This function adds an integer identifier value and a material that is red based on the number being 1 or blue based on the number being 0.
Added Chip prefab to prefabs folder.  Used in MySceneManager to instantiate new chips.

Added starter code that creates 5 default chips in a row to demonstrate that the chip set up that I've created works properly.  Uses random number generation of either 0 or 1 to assign values and materials to the correct chips being created.